+++ google/cloud/apigeeregistry/quickstart/quickstart.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 //! [all]
-#include "google/cloud/apigeeregistry/v1/ EDIT HERE _client.h"
+#include "google/cloud/apigeeregistry/v1/registry_client.h"
 #include "google/cloud/location.h"
 #include <iostream>
 
@@ -26,10 +26,9 @@ int main(int argc, char* argv[]) try {
   auto const location = google::cloud::Location(argv[1], argv[2]);
 
   namespace apigeeregistry = ::google::cloud::apigeeregistry_v1;
-  auto client = apigeeregistry::ServiceClient(
-      apigeeregistry::MakeServiceConnection());  // EDIT HERE
-
-  for (auto r : client.List /*EDIT HERE*/ (location.FullName())) {
+  auto client =
+      apigeeregistry::RegistryClient(apigeeregistry::MakeRegistryConnection());
+  for (auto r : client.ListApis(location.FullName())) {
     if (!r) throw std::move(r).status();
     std::cout << r->DebugString() << "\n";
   }